requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyephem==9.99
pydantic==2.5.0
python-multipart==0.0.6
pytz==2023.3.post1
httpx==0.28.1